Model/XAI/LIME/requirements.txt
torch>=1.13
torchvision>=0.14
lime>=0.2.0.1
numpy>=1.21
pillow>=9.0
scikit-image>=0.19
matplotlib>=3.5
scikit-learn>=1.1
